# openapi.yaml
openapi: 3.0.3
info:
  title: Gateway Zero API
  description: |
    Gateway Zero 是一个基于 Rust 和 Axum 框架开发的 AI API 网关系统。
    它提供统一的 OpenAI 兼容接口，支持多个 AI 提供商（OpenAI、Anthropic、智谱 AI 等），
    并具备负载均衡、令牌管理、用量统计等企业级功能。
  version: 0.1.0
  contact:
    name: Gateway Zero Team
servers:
  - url: http://localhost:8080
    description: 本地开发服务器
  - url: https://api.example.com
    description: 生产环境服务器
components:
  securitySchemes:
    # Client Token：外部调用 /v1/* 的 API Token（与 Admin Identity 严格区分）
    ClientToken:
      type: apiKey
      in: header
      name: Authorization
      description: "Client Token 认证（格式: Bearer <token>），用于访问 `/v1/*`"

    # Admin Identity：管理员身份（可来自 JWT AccessToken / TUI Session bearer / Web Session Cookie）
    AdminTuiSessionToken:
      type: http
      scheme: bearer
      description: 管理员身份（Admin Identity）的 TUI Session Token（bearer，但非 JWT）

    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: gw_session
      description: 管理员身份（Admin Identity）的 Web Session Cookie（HttpOnly）

    AccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "AccessToken（JWT），格式: Bearer <jwt>；用于访问 `/auth/me`、`/auth/change-password`、`/me/*` 以及 superadmin-only 的 `/admin/*`、`/providers/*`"
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: 错误代码 (如 http_error, not_found, rate_limited 等)
        message:
          type: string
          description: 错误详细信息

    # 请求日志
...
    Model:
      type: object
      properties:
        id:
          type: string
          description: 模型ID
        object:
          type: string
          description: 对象类型
        created:
          type: integer
          format: int64
          description: 创建时间戳
        owned_by:
          type: string
          description: 所有者
      required:
        - id
        - object
        - created
        - owned_by

    # 模型列表响应
    ModelsResponse:
      type: object
      properties:
        object:
          type: string
          default: list
        data:
          type: array
          items:
            $ref: '#/components/schemas/Model'
      required:
        - object
        - data

    # 聊天补全请求
...
paths:
...
  /v1/models:
    get:
      summary: 获取模型列表
      description: 获取所有可用的模型列表（支持 Client Token；已登录管理员身份也可访问）
      operationId: listModels
      tags:
        - Models
      security:
        - ClientToken: []
        - AccessToken: []
        - AdminTuiSessionToken: []
        - AdminSessionCookie: []
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelsResponse'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /models/{provider}:
    get:
      summary: 获取指定提供商模型列表
      description: 获取指定提供商的模型列表缓存；`refresh=true` 时拉取上游并返回（不落库）
      operationId: listProviderModels
      tags:
        - Models
      security:
        - ClientToken: []
        - AccessToken: []
        - AdminTuiSessionToken: []
        - AdminSessionCookie: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
          description: 提供商名称
        - name: refresh
          in: query
          schema:
            type: boolean
          description: 是否从上游刷新（返回但不写入缓存）
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelsResponse'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 提供商不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...